joystick: mapping axes
In this commit the `JoystickMapping` class has been split into
`JoystickAxes` and `JoystickButtons` to logically separate logically
separate enteties. One of the axes index is not correct accordingo
to the official ROS documentation. Tested with ros package `joy` and
`joy_linux`.

In this commit the axes has been mapped using a playstation 4
controller. Button mapping have been mapped from the official
documentation but not yet tested with hardware.