joystick: first working version
- imports: fix inconsistencies
- logging: logs new task on change not continiously
- fix/emulator: does no longer return NaN for yaw
- emulator: tuning
- static analysis: add mypy config
- type hinting: add more type hints, prefer ArrayLike over ndarray
- fix: bug where setting a new task immediatly resets it to default

There is currently a bug where button mappings have to be hard coded on
the enum class. I don't know why that is the case, but this is not needed
on the JoystickAxes class. Current workaround has been to set those
values on the class and not accept them from any other sources.