fix(joystick): requested thrust is not fully acutuated
Without this patch the produced command is not fully utilized. Rather the
produced command is at most 0.5. This could mean that the command is
being split across two azimuth thrusters or that there is some other
problem. Since tau symbolizes requested force in newton we then should
modify multiply joy with this scaling factor such that both thrusters
are taken into accout. I don't like this aproach because it's not
generic enough and leaves two random scalars in the code.